requirements.txt
streamlit
pillow
torch 
torchvision
tensorflow
numpy
tensorflow_addons
